requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
django-cors-headers==3.4.0
drf_yasg==1.17.1
gevent==20.6.2
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
dj_database_url
django-heroku
kombu==4.6.11
MarkupSafe==1.1.1
oauth2-provider==0.0
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
twilio==6.0.0rc13
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0